ISSUE #3879 - extract LabelButton from Button and simplify the latter
