+++ frontend/src/v5/ui/controls/button/button.styles.ts
@@ -16,116 +16,10 @@
  */
 
 import MuiButtonBase from '@mui/material/Button';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 export const MuiButton = styled(MuiButtonBase)`
 	white-space: nowrap;
-`;
-
-export const labelButtonPrimaryStyles = css`
-	color: ${({ theme }) => theme.palette.primary.contrast};
-	background-color: ${({ theme }) => theme.palette.primary.main};
-
-	&.Mui-focusVisible {
-		background-color: ${({ theme }) => theme.palette.primary.main};
-	}
-
-	&:hover {
-		background-color: ${({ theme }) => theme.palette.primary.dark};
-		text-decoration-line: none;
-	}
-
-	&:active {
-		background-color: ${({ theme }) => theme.palette.primary.darkest};
-	}
-`;
-
-export const labelButtonSecondaryStyles = css`
-	color: ${({ theme }) => theme.palette.tertiary.main};
-	background-color: ${({ theme }) => theme.palette.tertiary.lightest};
-
-	&:hover, &:active {
-		background-color: ${({ theme }) => theme.palette.tertiary.lighter};
-		text-decoration-line: none;
-	}
-`;
-
-export const labelOutlinedButtonPrimaryStyles = css`
-	color: ${({ theme }) => theme.palette.primary.main};
-	background-color: ${({ theme }) => theme.palette.primary.contrast};
-	border: 1px solid ${({ theme }) => theme.palette.primary.main};
-
-	&:hover {
-		color: ${({ theme }) => theme.palette.primary.contrast};
-		border-color: ${({ theme }) => theme.palette.primary.dark};
-	}
-
-	&:active {
-		border-color: ${({ theme }) => theme.palette.primary.darkest};
-	}
-
-	&.Mui-focusVisible {
-		color: ${({ theme }) => theme.palette.primary.main};
-		background-color: ${({ theme }) => theme.palette.primary.contast};
-	}
-
-	&:disabled {
-		border-color: ${({ theme }) => theme.palette.base.lightest};
-	}
-`;
-
-export const labelOutlinedButtonSecondaryStyles = css`
-	color: ${({ theme }) => theme.palette.tertiary.main};
-	background-color: ${({ theme }) => theme.palette.primary.contrast};
-	border: 1px solid ${({ theme }) => theme.palette.tertiary.main};
-
-	:active {
-		border-color: ${({ theme }) => theme.palette.tertiary.dark};
-	}
-
-	&.Mui-focusVisible {
-		background-color: ${({ theme }) => theme.palette.tertiary.lightest};
-	}
-
-	&:disabled {
-		border-color: ${({ theme }) => theme.palette.base.lightest};
-	}
-`;
-
-export const LabelButton = styled(MuiButton)<{ color?: string, outlined?: any }>`
-	align-items: center;
-	padding: 8px 12px 8px 15px;
-
-	${({ theme }) => css`
-		&:disabled {
-			background-color: ${theme.palette.base.lightest};
-			color: ${theme.palette.primary.contrast};
-		}
-
-		&.Mui-focusVisible {
-			box-shadow: ${theme.palette.shadows.level_5};
-		}
-	`}
-
-	${({ color }) => {
-		if (color === 'primary') {
-			return labelButtonPrimaryStyles;
-		}
-		if (color === 'secondary') {
-			return labelButtonSecondaryStyles;
-		}
-		return '';
-	}}
-
-	${({ color, outlined }) => {
-		if (color === 'primary' && outlined) {
-			return labelOutlinedButtonPrimaryStyles;
-		}
-		if (color === 'secondary' && outlined) {
-			return labelOutlinedButtonSecondaryStyles;
-		}
-		return '';
-	}}
 `;
 
 export const ErrorButton = styled(MuiButton).attrs({
