+++ frontend/src/v5/ui/controls/button/labelButton/labelButton.component.tsx
@@ -1,5 +1,5 @@
 /**
- *  Copyright (C) 2021 3D Repo Ltd
+ *  Copyright (C) 2023 3D Repo Ltd
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as
@@ -15,5 +15,14 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { Button } from './button.component';
-export { LabelButton } from './labelButton/labelButton.component';
+import { Typography } from '@/v5/ui/controls/typography';
+import { ButtonProps } from '@mui/material/Button';
+import { Button } from './labelButton.styles';
+
+type LabelButtonProps = Omit<ButtonProps, 'variant' | 'color'>;
+
+export const LabelButton = ({ children, ...props }: LabelButtonProps) => (
+	<Button {...props}>
+		<Typography variant="kicker">{children}</Typography>
+	</Button>
+);
