+++ frontend/src/v5/ui/v4Adapter/dialogs/generalDialogStyling.overrides.ts
@@ -16,7 +16,7 @@
  */
 import { css } from 'styled-components';
 import { DialogTitle } from '@/v4/routes/components/dialogContainer/components/dialog/dialog.styles';
-import { labelButtonPrimaryStyles } from '@controls/button/button.styles';
+import { primaryButtonStyling } from '@/v5/ui/v4Adapter/resuableOverrides.styles';
 import { Footer as InvitationsListFooter } from '@/v4/routes/components/invitationsDialog/invitationsDialog.styles';
 import { Footer as NewInviteFooter } from '@/v4/routes/components/invitationDialog/invitationDialog.styles';
 
@@ -69,7 +69,7 @@ const buttonContainerStyling = css`
 		background-color: ${({ theme }) => theme.palette.tertiary.lightest};
 
 		.MuiButtonBase-root.MuiButtonBase-root:last-child {
-			${labelButtonPrimaryStyles}
+			${primaryButtonStyling}
 			line-height: 1;
 
 			&, &:hover, &:active {
